Use IndexMap for storing trie node children
Currently, search order in the Trie is arbitrary due to
usage of HashMap for storing nodes. This is not practical
when trying to produce different search orders by shuffling
the dictionary.

This commit replaces usages of HashMap in Trie by an
IndexMap. IndexMap iteration order corresponds to insertion
order.

Also, the IndexMap used in Trie is configured to use the
FxHasher through the definition of type FxIndexMap, just
like HashMap before with the usage of FxHashMap.

Performance has not been assessed yet, but no penalty is
expected after this change.